refactor: Move pkgs: storage/level/local/writer/sourcenode -> storage/level/local/source
