+++ internal/pkg/service/stream/storage/level/local/source/writechain/closer.go
